modules/vpc/src/main/java/com/ibm/cloud/is/vpc/v1/model/ListSecurityGroupTargetsOptions.java
/*
 * (C) Copyright IBM Corp. 2021.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
 * the License. You may obtain a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
 * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 * specific language governing permissions and limitations under the License.
 */
package com.ibm.cloud.is.vpc.v1.model;

import com.ibm.cloud.sdk.core.service.model.GenericModel;

/**
 * The listSecurityGroupTargets options.
 */
public class ListSecurityGroupTargetsOptions extends GenericModel {

  protected String securityGroupId;
  protected String start;
  protected Long limit;

  /**
   * Builder.
   */
  public static class Builder {
    private String securityGroupId;
    private String start;
    private Long limit;

    private Builder(ListSecurityGroupTargetsOptions listSecurityGroupTargetsOptions) {
      this.securityGroupId = listSecurityGroupTargetsOptions.securityGroupId;
      this.start = listSecurityGroupTargetsOptions.start;
      this.limit = listSecurityGroupTargetsOptions.limit;
    }

    /**
     * Instantiates a new builder.
     */
    public Builder() {
    }

    /**
     * Instantiates a new builder with required properties.
     *
     * @param securityGroupId the securityGroupId
     */
    public Builder(String securityGroupId) {
      this.securityGroupId = securityGroupId;
    }

    /**
     * Builds a ListSecurityGroupTargetsOptions.
     *
     * @return the new ListSecurityGroupTargetsOptions instance
     */
    public ListSecurityGroupTargetsOptions build() {
      return new ListSecurityGroupTargetsOptions(this);
    }

    /**
     * Set the securityGroupId.
     *
     * @param securityGroupId the securityGroupId
     * @return the ListSecurityGroupTargetsOptions builder
     */
    public Builder securityGroupId(String securityGroupId) {
      this.securityGroupId = securityGroupId;
      return this;
    }

    /**
     * Set the start.
     *
     * @param start the start
     * @return the ListSecurityGroupTargetsOptions builder
     */
    public Builder start(String start) {
      this.start = start;
      return this;
    }

    /**
     * Set the limit.
     *
     * @param limit the limit
     * @return the ListSecurityGroupTargetsOptions builder
     */
    public Builder limit(long limit) {
      this.limit = limit;
      return this;
    }
  }

  protected ListSecurityGroupTargetsOptions(Builder builder) {
    com.ibm.cloud.sdk.core.util.Validator.notEmpty(builder.securityGroupId,
      "securityGroupId cannot be empty");
    securityGroupId = builder.securityGroupId;
    start = builder.start;
    limit = builder.limit;
  }

  /**
   * New builder.
   *
   * @return a ListSecurityGroupTargetsOptions builder
   */
  public Builder newBuilder() {
    return new Builder(this);
  }

  /**
   * Gets the securityGroupId.
   *
   * The security group identifier.
   *
   * @return the securityGroupId
   */
  public String securityGroupId() {
    return securityGroupId;
  }

  /**
   * Gets the start.
   *
   * A server-supplied token determining what resource to start the page on.
   *
   * @return the start
   */
  public String start() {
    return start;
  }

  /**
   * Gets the limit.
   *
   * The number of resources to return on a page.
   *
   * @return the limit
   */
  public Long limit() {
    return limit;
  }
}

